Aggregation Method: AverageWAgg

Calculate one of several types of averaged best estimates.

AverageWAgg(
  expert_judgements,
  type = "ArMean",
  name = NULL,
  placeholder = FALSE,
  percent_toggle = FALSE
)

Arguments

expert_judgements

A dataframe in the format of data_ratings.

type

One of "ArMean", "Median", "GeoMean", "LOArMean", or "ProbitArMean".

name

Name for aggregation method. Defaults to type unless specified.

placeholder

Toggle the output of the aggregation method to impute placeholder data.

percent_toggle

Change the values to probabilities. Default is FALSE.

Value

A tibble of confidence scores cs for each paper_id.

Details

This function returns the average, median and transformed averages of best-estimate judgements for each claim.

type may be one of the following:

ArMean: Arithmetic mean of the best estimates \[\hat{p}_c\left(ArMean \right ) = \frac{1}{N}\sum_{i=1}^N B_{i,c}\] Median: Median of the best estimates \[\hat{p}_c \left(\text{median} \right) = \text{median} { B^i_c}_{i=1,...,N}\] GeoMean: Geometric mean of the best estimates \[GeoMean_{c}= \left(\prod_{i=1}^N B_{i,c}\right)^{\frac{1}{N}}\] LOArMean: Arithmetic mean of the log odds transformed best estimates \[LogOdds_{i,c}= \frac{1}{N} \sum_{i=1}^N log\left( \frac{B_{i,c}}{1-B_{i,c}}\right)\] The average log odds estimate is then back transformed to give a final group estimate: \[\hat{p}_c\left( LOArMean \right) = \frac{e^{LogOdds_{i,c}}}{1+e^{LogOdds_{i,c}}}\] ProbitArMean: Arithmetic mean of the probit transformed best estimates \[Probit_{c}= \frac{1}{N} \sum_{i=1}^N \Phi^{-1}\left( B_{i,c}\right)\] The average probit estimate is then back transformed to give a final group estimate: \[\hat{p}_c\left(ProbitArMean \right) = \Phi\left({Probit_{c}}\right)\]

Examples

if (FALSE) AverageWAgg(data_ratings)