Aggregation Method: LinearWAgg

Calculate one of several types of linear-weighted best estimates.

LinearWAgg(
  expert_judgements,
  type = "DistLimitWAgg",
  weights = NULL,
  name = NULL,
  placeholder = FALSE,
  percent_toggle = FALSE,
  flag_loarmean = FALSE
)

Arguments

expert_judgements

A dataframe in the format of data_ratings.

type

One of "Judgement", "Participant", "DistLimitWAgg", "GranWAgg", or "OutWAgg".

weights

(Optional) A two column dataframe (user_name and weight) for type = "Participant" or a three two column dataframe (paper_id', 'user_name and weight) for type = "Judgement"

name

Name for aggregation method. Defaults to type unless specified.

placeholder

Toggle the output of the aggregation method to impute placeholder data.

percent_toggle

Change the values to probabilities. Default is FALSE.

flag_loarmean

A toggle to impute log mean (defaults FALSE).

Value

A tibble of confidence scores cs for each paper_id.

Details

This function returns weighted linear combinations of the best-estimate judgements for each claim.

type may be one of the following:

Judgement: Weighted by user-supplied weights at the judgement level \[\hat{p}_c\left( JudgementWeights \right) = \sum_{i=1}^N \tilde{w}\_judgement_{i,c}B_{i,c}\]

Participant: Weighted by user-supplied weights at the participant level \[\hat{p}_c\left( ParticipantWeights \right) = \sum_{i=1}^N \tilde{w}\_participant_{i}B_{i,c}\]

DistLimitWAgg: Weighted by the distance of the best estimate from the closest certainty limit. Giving greater weight to best estimates that are closer to certainty limits may be beneficial. \[w\_distLimit_{i,c} = \max \left(B_{i,c}, 1-B_{i,c}\right)\] \[\hat{p}_c\left( DistLimitWAgg \right) = \sum_{i=1}^N \tilde{w}\_distLimit_{i,c}B_{i,c}\]

GranWAgg: Weighted by the granularity of best estimates

Individuals are weighted by whether or not their best estimates are more granular than a level of 0.05 (i.e., not a multiple of 0.05). \[w\_gran_{i} = \frac{1}{C} \sum_{d=1}^C \left\lceil{\frac{B_{i,d}} {0.05}-\left\lfloor{\frac{B_{i,d}}{0.05}}\right\rfloor}\right\rceil,\]

where \(\lfloor{\ }\rfloor\) and \(\lceil{\ }\rceil\) are the mathematical floor and ceiling functions respectively. \[\hat{p}_c\left( GranWAgg \right) = \sum_{i=1}^N \tilde{w}\_gran_{i} B_{i,c}\]

OutWAgg: Down weighting outliers

This method down-weights outliers by using the differences from the central tendency (median) of an individual's best estimates. \[d_{i,c} = \left(median{{B_{i,c}}_{_{i=1,...,N}}} - B_{i,c}\right)^2\] \[w\_out_{i} = 1 - \frac{d_{i,c}}{\max({d_c})})\] \[\hat{p}_c\left( OutWAgg \right) = \sum_{i=1}^N \tilde{w}\_out_{i}B_{i,c}\]

Examples

if (FALSE) LinearWAgg(data_ratings)