Confidence Score Heat Map

Confidence scores displayed on a colour spectrum across generated methods and assessed claims, split into predicted replication outcomes (TRUE/FALSE). White indicative of around .5 with higher predicted confidence scores more blue (>.5) and lower more red (<.5). Each predicted replication outcome is then split into the group type of the underlying statistical characteristic for each aggregation method (non-weighted linear, weighted linear & Bayesian).

confidence_score_heatmap(
  confidence_scores = NULL,
  data_outcomes = NULL,
  x_label = NULL
)

Arguments

confidence_scores

A data frame of confidence scores generated from the aggregation methods in the form of data_confidence_scores. Defaults to data_confidence_scores if no argument supplied.

data_outcomes

A data frame of unique claims and the associated binary outcome in the form of data_outcomes. If no argument supplied then defaults to data_outcomes supplied within package.

x_label

Bottom x axis label name or ID. Default is blank.

Value

Plot in viewer

Examples

if (FALSE) confidencescore_heatmap(data_confidence_scores, data_outcomes)