Weighting method: Asymmetry of intervals

Calculates weights by asymmetry of intervals

weight_asym(expert_judgements)

Arguments

expert_judgements

the long tibble exported from the preprocess_judgements function.

Value

A tibble of confidence scores (value) for the three-point best estimate (element), including weighting values for the magnitude of asymmetry between lower, best, and upper estimates.

Details

This function is used inside IntervalWAgg to calculate the weights for the aggregation type "AsymWAgg", "IndIntAsymWAgg" and "KitchSinkWAgg". Pre-processed expert judgements (long format) are first converted to wide format then weighted by: \[w\_asym_{i,c}= \begin{cases} 1 - 2 \cdot \frac{U_{i,c}-B_{i,c}}{U_{i,c}-L_{i,c}}, \text{for}\ B_{i,c} \geq \frac{U_{i,c}-L_{i,c}}{2}+L_{i,c}\cr 1 - 2 \cdot \frac{B_{i,c}-L_{i,c}}{U_{i,c}-L_{i,c}}, \text{otherwise} \end{cases}\]

Data is converted back to long format, with only the weighted best estimates retained.

Examples

weight_asym(preprocess_judgements(data_ratings))
#> 
#> -- Pre-Processing Options --
#> 
#> i Round Filter: TRUE
#> i Three Point Filter: TRUE
#> i Percent Toggle: FALSE
#> # A tibble: 625 x 8
#>    round   paper_id user_name     ul weight_obs agg_weight element         value
#>    <chr>   <chr>    <chr>      <dbl>      <dbl>      <dbl> <chr>           <dbl>
#>  1 round_2 100      7l8m7dmjdb    15      0.2        0.2   three_point_be~    39
#>  2 round_2 100      bvpgcjm09w    40      0          0     three_point_be~    70
#>  3 round_2 100      mnlww4qlq3    40      0.25       0.25  three_point_be~    65
#>  4 round_2 100      r7kn2x3mnj    35      0.6        0.6   three_point_be~    88
#>  5 round_2 100      lvr6dwbqag    40      0          0     three_point_be~    50
#>  6 round_2 100      1uvpofirab    25      0.2        0.2   three_point_be~    60
#>  7 round_2 100      200frs5sdp    25      0.2        0.2   three_point_be~    80
#>  8 round_2 100      stamxz9tm3    30      0          0     three_point_be~    80
#>  9 round_2 100      o8s51ahsct    30      0.333      0.333 three_point_be~    80
#> 10 round_2 100      1xfrz0c1ir    25      0.6        0.6   three_point_be~    85
#> # ... with 615 more rows